split kmergraph files in two
